Removed unnecessary methods, split tracking.cpp, some comments
