+++ CMakeLists.txt
@@ -10,6 +10,6 @@ link_directories(/home/flo/Workspace/darknet)
 
 
 add_executable(Panorama2Cubemap main.cpp opencv_detect.cpp
-        cubetransform.cpp cubetransform.h detect.cpp detect.h tracking.cpp opencv_tracking.cpp opencv_tracking.h PanoramaTracking.cpp PanoramaTracking.h PanoramaTrackingImplementations.cpp PanoramaTrackingImplementations.h TrackingHelpers.h testMaskRCNN.cpp testMaskRCNN.h)
+        cubetransform.cpp cubetransform.h detect.cpp detect.h tracking.cpp opencv_tracking.cpp opencv_tracking.h PanoramaTracking.cpp PanoramaTracking.h PanoramaTrackingImplementations.cpp PanoramaTrackingImplementations.h TrackingHelpers.h testMaskRCNN.cpp testMaskRCNN.h TrackingHelpers.cpp)
 
 target_link_libraries(Panorama2Cubemap -lm -lpthread -ldarknet ${OpenCV_LIBS})
