+++ tracking.h
@@ -51,7 +51,6 @@ public:
 
     virtual ~RegionTracker();
 
-
     Mat matCurrentFrame;
 
     FILE * roiData;
@@ -123,6 +122,9 @@ protected:
 
 };
 
+/*
+ * Some Helper Functions.
+ */
 void textAboveRect(Mat frame, Rect rect, string text);
 void histFromRect(Mat const & input, Rect const & rect, Mat & output);
 /* Color difference as described by https://en.wikipedia.org/wiki/Color_difference#CIE94
@@ -136,6 +138,6 @@ double deltaECIE94(unsigned char L1, char  a1, char b1, unsigned char L2, char a
  * Labels has one Row and frame.cols * frame.rows columns, representing the frames pixel like this : frame[row, col] = labels[row * frame.cols + col]
  */
 Mat helperBGRKMean(Mat const &frame, int clusterCount, Mat &labels, Mat &centers);
-
+bool playerInRegionVector(FootballPlayer * fp, vector<Region> const & vr);
 // void optimizeWeightSelection(int rows, int cols, double * weightMatrix, int * selected)
 #endif //PANORAMA2CUBEMAP_TRACKING_H
